fix: Update Docusaurus version and Node.js requirement
Update Docusaurus version from 2.4.1 to 3.0.1 and Node.js requirement from 16.14 or above to 18.0 or above.
